% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_overdispersion}
\alias{extract_overdispersion}
\title{Extract Overdispersion Vector}
\usage{
extract_overdispersion(seurat_list)
}
\arguments{
\item{seurat_list}{A list of Seurat objects.}
}
\value{
A named numeric vector of overdispersion values. Returns \code{NULL} if
"scaling_factor" is not found in the feature metadata.
}
\description{
A wrapper function to extract the default overdispersion vector
("scaling_factor") from a list of Seurat objects.
}
\details{
This function is a convenient helper that uses
\code{\link{extract_feature_vector}} to retrieve overdispersion estimates.
}
